Correct compiler warnings

+++ src/include/statements.h
@@ -64,7 +64,7 @@ typedef struct Value {
     union {
         uint64_t const_unsigned;
         int64_t  const_signed;
-        uint64_t local_id;
+        unsigned local_id;
     };
 } Value;
 
@@ -82,22 +82,22 @@ typedef struct Operation {
     Statement statement;
     uint8_t type;
     uint8_t var_type;
-    uint64_t dest; // ID of destination variable.
-    uint64_t lhs; // ID of source varible; may not be a constant.
+    unsigned dest; // ID of destination variable.
+    unsigned lhs; // ID of source varible; may not be a constant.
     Value rhs; // Either a constant or a local ID. Ignored unless using binops.
 } Operation;
 
 typedef struct Read {
     Statement statement;
     uint8_t var_type;
-    uint64_t dest;
+    unsigned dest;
     char* src;
 } Read;
 
 typedef struct Write {
     Statement statement;
     char* dest;
-    uint64_t src;
+    unsigned src;
 } Write;
 
 typedef struct Jump {
